# zafora-backend/src/workplace/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Zafora Holding API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: sector
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: country
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
                  total:
                    type: integer
                required: [projects, total]
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all consulting services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: object
                properties:
                  services:
                    type: array
                    items:
                      $ref: "#/components/schemas/Service"
                required: [services]
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceBody"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
...
  /content/stats:
    get:
      operationId: listContentStats
      tags: [content]
      summary: List editable site stats
      responses:
        "200":
          description: List of stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  stats:
                    type: array
                    items:
                      $ref: "#/components/schemas/ContentStat"
                required: [stats]
    post:
      operationId: createContentStat
      tags: [content]
      summary: Create a content stat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContentStatBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentStat"
  /content/stats/{id}:
    patch:
      operationId: updateContentStat
      tags: [content]
      summary: Update a content stat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContentStatBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentStat"
    delete:
      operationId: deleteContentStat
      tags: [content]
      summary: Delete a content stat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /content/methodology:
    get:
      operationId: listMethodologySteps
      tags: [content]
      summary: List methodology steps
      responses:
        "200":
          description: List of steps
          content:
            application/json:
              schema:
                type: object
                properties:
                  steps:
                    type: array
                    items:
                      $ref: "#/components/schemas/MethodologyStep"
                required: [steps]
    post:
      operationId: createMethodologyStep
      tags: [content]
      summary: Create a methodology step
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMethodologyStepBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MethodologyStep"
  /content/methodology/{id}:
    patch:
      operationId: updateMethodologyStep
      tags: [content]
      summary: Update a methodology step
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMethodologyStepBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MethodologyStep"
    delete:
      operationId: deleteMethodologyStep
      tags: [content]
      summary: Delete a methodology step
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /content/settings/{key}:
    get:
      operationId: getSiteSettings
      tags: [content]
      summary: Get site settings by key
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Settings value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSetting"
    patch:
      operationId: updateSiteSettings
      tags: [content]
      summary: Update site settings by key
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                value:
                  type: string
              required: [value]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSetting"
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sector:
          type: string
        country:
          type: string
        region:
          type: string
          nullable: true
        fundingStatus:
          type: string
        estimatedValue:
          type: string
        zaforaRole:
          type: string
        partnerNeed:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        interestCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, sector, country, fundingStatus, estimatedValue, zaforaRole, interestCount, createdAt]
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        sector:
          type: string
        country:
          type: string
        region:
          type: string
          nullable: true
        fundingStatus:
          type: string
        estimatedValue:
          type: string
        zaforaRole:
          type: string
        partnerNeed:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
      required: [name, sector, country, fundingStatus, estimatedValue, zaforaRole]
...
    Service:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        description:
          type: string
        bullets:
          type: array
          items:
            type: string
        imageUrl:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        displayOrder:
          type: integer
        visible:
          type: boolean
      required: [id, name, icon, description, bullets, displayOrder, visible]
    CreateServiceBody:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        description:
          type: string
        bullets:
          type: array
          items:
            type: string
        imageUrl:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        displayOrder:
          type: integer
        visible:
          type: boolean
      required: [name, description, bullets]
...
    ContentStat:
      type: object
      properties:
        id:
          type: integer
        label:
          type: string
        value:
          type: string
        suffix:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        iconName:
          type: string
          nullable: true
        displayOrder:
          type: integer
        visible:
          type: boolean
      required: [id, label, value, displayOrder, visible]
    CreateContentStatBody:
      type: object
      properties:
        label:
          type: string
        value:
          type: string
        suffix:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        iconName:
          type: string
          nullable: true
        displayOrder:
          type: integer
        visible:
          type: boolean
      required: [label, value]
    UpdateContentStatBody:
      type: object
      properties:
        label:
          type: string
        value:
          type: string
        suffix:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        iconName:
          type: string
          nullable: true
        displayOrder:
          type: integer
        visible:
          type: boolean
    MethodologyStep:
      type: object
      properties:
        id:
          type: integer
        stepNumber:
          type: integer
        title:
          type: string
        description:
          type: string
        iconName:
          type: string
          nullable: true
        displayOrder:
          type: integer
        visible:
          type: boolean
      required: [id, stepNumber, title, description, displayOrder, visible]
    CreateMethodologyStepBody:
      type: object
      properties:
        stepNumber:
          type: integer
        title:
          type: string
        description:
          type: string
        iconName:
          type: string
          nullable: true
        displayOrder:
          type: integer
        visible:
          type: boolean
      required: [stepNumber, title, description]
    UpdateMethodologyStepBody:
      type: object
      properties:
        stepNumber:
          type: integer
        title:
          type: string
        description:
          type: string
        iconName:
          type: string
          nullable: true
        displayOrder:
          type: integer
        visible:
          type: boolean
    SiteSetting:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
        value:
          type: string
        updatedAt:
          type: string
      required: [id, key, value, updatedAt]
...